+++ src/Typin/Typin.Core/Pipeline.cs
@@ -6,5 +6,5 @@
     /// Represents an async continuation for the next task to execute in the pipeline.
     /// </summary>
     /// <returns>Awaitable task</returns>
-    public delegate Task CommandPipelineHandlerDelegate();
+    public delegate ValueTask CommandPipelineHandlerDelegate();
 }