+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Core middleware execution order has changed: `ResolveCommandSchema` -> `ResolveCommandInstance` -> [User middlewares] -> `HandleSpecialOptions` -> `ExecuteCommand`).
 - Removed unnecessary casts to `CliContext` from `ICliContext`.
 - Removed `IDirective.ContinueExecution`, modified `IDirective`, and added `IPipelinedDirective`.
+- `CommandPipelineHandlerDelegate` now uses `ValueTask` instead of a `Task`.
 
 ### v2.1.1 (18-Oct-2020)